Working Procedure
#Working with Dialog Flow APi
Go to Dialogflow and click on console. Signup using your google account
Grant permissions and Accept the terms and conditions agreement.	
Click on create new agent.
Give a name to your agent and click save.
Now click on the small talks tab on the left menu and enable it.

#Creating Client_Access Token

Give answers for all the question to train your agent and save it. You can give multiple answers for a single question.
Click on the gear icon, copy your Client Access Token from their and paste it in the code.

#Code Explanation

We will add a new function in our existing code to establish a connection between our bot and dialogflow. 
 *Connect to dialogflow
    def apiaiCon(self):
        self.CLIENT_ACCESS_TOKEN = "Your Client Access Token"
        self.ai = apiai.ApiAI(self.CLIENT_ACCESS_TOKEN)
        self.request = self.ai.text_request()
        self.request.lang = 'de' #Default : English
 
On receiving new message, we will request a reply for that message from dialogflow which will send us a json object in return. 

We will convert the json object into a python list and store the reply in a variable reply. If you are using python 3.6 you can skip the decoding part. 
        # Convert json obect to a list. For python 3.6 replace json.load(reader(response)) with json.load(response)
        reader = codecs.getdecoder("utf-8")
        obj = json.load(reader(response))
 
        # Get reply from the list
        reply = obj['result']['fulfillment']['speech']
and that’s it! You have a fully working chatbot for your personal Facebook profile.
#How It Works
We will login from a Facebook account using a script and wait and listen for new incoming messages. Once we receive a new message, we will get the text of that message, feed it into a trained model which will do the magic for us using natural language processing and machine learning   and return a response. We will then send that response to the person who sent you the message on Facebook.